Update OpenConfig public models, ygot, and ondatra dependencies (#194)

+++ feature/bgp/gracefulrestart/gracefulrestart.go
@@ -58,7 +58,7 @@ func (gr *GracefulRestart) WithRestartTime(secs time.Duration) *GracefulRestart 
 
 // WithStaleRoutesTime sets the stale routes time for graceful restart feature.
 func (gr *GracefulRestart) WithStaleRoutesTime(secs time.Duration) *GracefulRestart {
-	rt := ygot.Float64(secs.Seconds())
+	rt := ygot.Uint16(uint16(secs.Seconds()))
 	gr.noc.StaleRoutesTime = rt
 	gr.poc.StaleRoutesTime = rt
 	gr.goc.StaleRoutesTime = rt
